requirements.txt
asgiref==3.7.2
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dj-database-url==2.1.0
Django==5.0.3
gunicorn==21.2.0
h11==0.14.0
idna==3.6
outcome==1.3.0.post0
packaging==24.0
pillow==10.2.0
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.31.0
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.4.4
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.28.1
webdriver-manager==4.0.1
whitenoise==6.6.0
wsproto==1.2.0
